OLMIS-1713: Added submit endpoint for POD
In the document a user should be able to submit a Proof of delivery
when all data are provided. This endpoint was not present in the related
Controller class.

The user has ability to save any proof of delivery. There is no validation
on create/update endpoints. The object is validated when the user execute
the submit endpoint. If there will be any validation error the user will
get appropriate error message with details about what in the POD is not
valid.