OLMIS-1763: Added status parameter to order search endpoint
A user should be able to find orders with the given status.

The search endpoint will accept a new query parameter `status` that
could have any value but if the value does not match a value from
OrderStatus enum class the endpoint will return 400 error code with
appropriate error message.