OLMIS-1566 Update order search endpoint not to require warehouse
Do not need supplying facility to do the view order check since we are automatically filtering the orders by right anyways. Need to check PODS_MANAGE right in addition to ORDERS_VIEW right, since the user could have either one to get the list of orders. No need for 403 Forbidden because if the user does not have permission, the orders list should be empty.